Update card component codes
코드 스타일 룰에 맞춰 styled component 정의 코드를 index에 통합
price container의 컴포넌트 단위가 작다고 판단하여 card component에 통합
